+++ module/io/dap_events_socket.c
@@ -220,8 +220,8 @@ int dap_events_socket_queue_data_send(dap_events_socket_t *a_es, const void *a_d
     }
     return InterlockedPushEntrySList((PSLIST_HEADER)a_es->buf_out, &(l_entry->entry))
         ? a_size : PostQueuedCompletionStatus(a_es->context->iocp, a_size, (ULONG_PTR)a_es, NULL)
-            ? a_size : ( DAP_ALFREE(l_entry), log_it(L_ERROR, "Enqueue into es "DAP_FORMAT_ESOCKET_UUID" failed, errno %d",
-                                                              a_es->uuid, GetLastError()), 0 );
+            ? a_size : ( DAP_ALFREE(l_entry), log_it(L_ERROR, "Enqueue into es "DAP_FORMAT_ESOCKET_UUID" failed, errno %lu",
+                                                              a_es->uuid, (unsigned long)GetLastError()), 0 );
 }
 #endif
 
@@ -419,7 +419,7 @@ void dap_events_socket_reassign_between_workers(dap_worker_t *a_worker_old, dap_
     ol->ol.Internal = a_es_uuid;
     ol->op = io_call;
     if ( !PostQueuedCompletionStatus(a_worker_old->context->iocp, 0, (ULONG_PTR)s_es_reassign, (OVERLAPPED*)ol) ) {
-        log_it(L_ERROR, "Can't reassign es % " DAP_UINT64_FORMAT_x ",  error %d", a_es_uuid, GetLastError());
+        log_it(L_ERROR, "Can't reassign es % " DAP_UINT64_FORMAT_x ",  error %lu", a_es_uuid, (unsigned long)GetLastError());
         dap_overlapped_free(ol);
     }
     return;
@@ -1551,8 +1551,8 @@ void dap_events_socket_remove_and_delete(dap_worker_t *a_worker, dap_events_sock
     ol->ol.OffsetHigh = 1;
     ol->op = io_call;
     if ( !PostQueuedCompletionStatus(a_worker->context->iocp, 0, (ULONG_PTR)s_es_set_flag, (OVERLAPPED*)ol) ) {
-        log_it(L_ERROR, "Can't schedule deletion of %"DAP_UINT64_FORMAT_U" in context #%d, error %d",
-               a_es_uuid, a_worker->context->id, GetLastError());
+        log_it(L_ERROR, "Can't schedule deletion of %"DAP_UINT64_FORMAT_U" in context #%d, error %lu",
+               a_es_uuid, a_worker->context->id, (unsigned long)GetLastError());
         dap_overlapped_free(ol);
     }
 #else
@@ -1594,8 +1594,8 @@ void dap_events_socket_set_readable(dap_worker_t *a_worker, dap_events_socket_uu
     ol->ol.OffsetHigh = (DWORD)a_is_ready;
     ol->op = io_call;
     if ( !PostQueuedCompletionStatus(a_worker->context->iocp, 0, (ULONG_PTR)s_es_set_flag, (OVERLAPPED*)ol) ) {
-        log_it(L_ERROR, "Can't schedule reading from %"DAP_UINT64_FORMAT_U" in context #%d, error %d",
-               a_es_uuid, a_worker->context->id, GetLastError());
+        log_it(L_ERROR, "Can't schedule reading from %"DAP_UINT64_FORMAT_U" in context #%d, error %lu",
+               a_es_uuid, a_worker->context->id, (unsigned long)GetLastError());
         dap_overlapped_free(ol);
     }
 #else
@@ -1638,8 +1638,8 @@ void dap_events_socket_set_writable(dap_worker_t *a_worker, dap_events_socket_uu
     ol->ol.OffsetHigh = (DWORD)a_is_ready;
     ol->op = io_call;
     if ( !PostQueuedCompletionStatus(a_worker->context->iocp, 0, (ULONG_PTR)s_es_set_flag, (OVERLAPPED*)ol) ) {
-        log_it(L_ERROR, "Can't schedule writing to %"DAP_UINT64_FORMAT_U" in context #%d, error %d",
-               a_es_uuid, a_worker->context->id, GetLastError());
+        log_it(L_ERROR, "Can't schedule writing to %"DAP_UINT64_FORMAT_U" in context #%d, error %lu",
+               a_es_uuid, a_worker->context->id, (unsigned long)GetLastError());
         dap_overlapped_free(ol);
     }
 #else
